pythonProject2/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
chromedriver-binary-auto==0.1
click==7.1.2
colorama==0.4.4
dotenv-python==0.0.1
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
selenium==3.141.0
toml==0.10.2
urllib3==1.26.3
